+++ tests/Extension/BuildTest.php
@@ -4,17 +4,17 @@ declare(strict_types=1);
 
 namespace Version\Tests\Extension;
 
-use Version\Extension\BaseExtension;
+use Version\Extension\Extension;
 use Version\Extension\Build;
 
-class BuildTest extends BaseExtensionTest
+class BuildTest extends ExtensionTest
 {
-    protected function createExtension($identifiers): BaseExtension
+    protected function createExtension($identifiers): Extension
     {
         if (is_string($identifiers)) {
-            return Build::fromIdentifiersString($identifiers);
+            return Build::fromString($identifiers);
         }
 
-        return Build::fromIdentifiers(...$identifiers);
+        return Build::from(...$identifiers);
     }
 }